+++ dao/src/main/java/by/pvt/medvedeva/education/dao/interfacesDAO/BaseDAO.java
@@ -22,7 +22,7 @@ public interface BaseDAO <T extends Pojo> {
      * @return
      * @throws DAOException
      */
-    T getById(Integer id) throws DAOException;
+    T getById(Long id) throws DAOException;
 
     /**
      * @param pojo
@@ -41,5 +41,5 @@ public interface BaseDAO <T extends Pojo> {
      * @param id
      * @throws DAOException
      */
-    void delete(Integer id) throws DAOException;
+    void delete(Long id) throws DAOException;
 }
